fix data layout for LLVM x86_64-unknown-none target.
Honstly I don't understand this, if the data layout has to be exactly
the same as the LLVM target default, what's the point of being able to
specify it? And, this field is mandatory!

Also my privious layout DOES NOT CONFLICT with the target default, it's
just a subset! And the compiler is complaining... WHAT?????