toolchain: replace xbuild with build-std
for a no_std build with custom target, we need to build the rust
compoments including core, alloc and compiler builtins. Previously we
do this with the cargo xbuild tool. however it has some bugs with newer
toolchains. The official build-std feature is not stable but looks
promising! Also we could reduce a lot of build dependencies.

References: https://github.com/rust-in-action/code/issues/14
Signed-off-by: Tianhao Wang <shrik3@mailbox.org>