multiboot: defines multiboot variables in high mem
previously I reserved some bytes in 32 bit data to store the mbinfo
address and magic so they can be accessed from 32 bit init asm code.
This is dangerous because they will be lost once we drop the lower
memory mapping for user progs. It's possible to convert the addresses to
high mem mapping with some pointer magic but that's rather ugly and
unreadable.

Signed-off-by: Tianhao Wang <shrik3@mailbox.org>